+++ src/interfaces/components/authorization/Oauth2ClientAuthorizationRequestInterface.php
@@ -2,8 +2,8 @@
 
 namespace rhertogh\Yii2Oauth2Server\interfaces\components\authorization;
 
+use rhertogh\Yii2Oauth2Server\interfaces\models\external\user\Oauth2UserInterface;
 use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2ClientInterface;
-use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2UserInterface;
 use rhertogh\Yii2Oauth2Server\Oauth2Module;
 use yii\base\Configurable;