+++ src/traits/models/Oauth2OidcUserIdentityTrait.php
@@ -3,7 +3,7 @@
 namespace rhertogh\Yii2Oauth2Server\traits\models;
 
 use rhertogh\Yii2Oauth2Server\interfaces\components\openidconnect\scope\Oauth2OidcClaimInterface;
-use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2OidcUserInterface;
+use rhertogh\Yii2Oauth2Server\interfaces\models\external\user\Oauth2OidcUserInterface;
 use rhertogh\Yii2Oauth2Server\Oauth2Module;
 use yii\base\InvalidArgumentException;
 use yii\helpers\ArrayHelper;