+++ src/traits/models/Oauth2UserIdentityTrait.php
@@ -2,7 +2,7 @@
 
 namespace rhertogh\Yii2Oauth2Server\traits\models;
 
-use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2UserInterface;
+use rhertogh\Yii2Oauth2Server\interfaces\models\external\user\Oauth2UserInterface;
 use rhertogh\Yii2Oauth2Server\Oauth2Module;
 use Yii;
 use yii\web\IdentityInterface;