+++ tests/_helpers/TestUserModel.php
@@ -2,7 +2,7 @@
 
 namespace Yii2Oauth2ServerTests\_helpers;
 
-use rhertogh\Yii2Oauth2Server\interfaces\models\Oauth2UserInterface;
+use rhertogh\Yii2Oauth2Server\interfaces\models\external\user\Oauth2UserInterface;
 use rhertogh\Yii2Oauth2Server\Oauth2Module;
 use rhertogh\Yii2Oauth2Server\traits\models\Oauth2UserIdentityTrait;
 use Yii;